Create seconds_from_time
A program that converts hours, minutes and seconds in seconds and a number provided, the last two digits represent the number of seconds, the third and fourth digits (from the right) represent the number of minutes and the remaining digits (from the fifth, counting from right to left) represent the number of hours. For example, 124927 means 12 hours, 49 minutes and 27 seconds.