Create Gasoline
Fortunately, the price of gasoline has been steadily falling. So, we can again fill up the tanks of our Porsches and BMWs without big bad conscience. But our memory is short. How long ago has the price of gasoline started to drop?

Task

Write a program that reads one sequence of numbers double , representing the price of gasoline, every day, and computes how many days ago the has price started to decline, more recently.

More precisely, the day when the price started to decline more recently is the first day the price dropped after the last day on which the price rose. Note that the price goes up on day d if the price on day d is greater than the price on day d -1; the price goes down on d if the price on the day d is less than the price on d -1. If, by hypothesis, today day x then yesterday is the day x -1, etc. If the price of gasoline began to descend on day y , then it started to descend x - y days ago.

If the price of gasoline has not gone down since the last time it went up, your program should write "void" (without the quotes). Ditto, if the data do not record any price increase.